+++ packages/web3-eth/test/unit/prepare_transaction_for_signing.test.ts
@@ -16,7 +16,6 @@ along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 
 import {
-	Common,
 	EthExecutionAPI,
 	HexString,
 	Web3NetAPI,
@@ -33,6 +32,7 @@ import {
 	AccessListEIP2930Transaction,
 	FeeMarketEIP1559Transaction,
 	Transaction,
+	Hardfork,
 } from 'web3-eth-accounts';
 import { prepareTransactionForSigning } from '../../src/utils/prepare_transaction_for_signing';
 import { validTransactions } from '../fixtures/prepare_transaction_for_signing';
@@ -70,7 +70,7 @@ describe('prepareTransactionForSigning', () => {
 
 				if (isNullish(tx.common)) {
 					if (options.web3Context.defaultCommon) {
-						const common = options.web3Context.defaultCommon as unknown as Common;
+						const common = options.web3Context.defaultCommon;
 						const chainId = common.customChain.chainId as string;
 						const networkId = common.customChain.networkId as string;
 						const name = common.customChain.name as string;
@@ -102,6 +102,153 @@ describe('prepareTransactionForSigning', () => {
 			expect(ethereumjsTx.common.chainName()).toBe('test');
 		});
 	});
+
+	it('should be able to read Hardfork from context.defaultHardfork', async () => {
+		const context = new Web3Context<EthExecutionAPI>({
+			provider: new HttpProvider('http://127.0.0.1'),
+			config: { defaultNetworkId: '0x9' },
+		});
+		context.defaultChain = 'mainnet';
+		context.defaultHardfork = Hardfork.Istanbul;
+
+		async function transactionBuilder<ReturnType = TransactionType>(options: {
+			transaction: TransactionType;
+			web3Context: Web3Context<EthExecutionAPI & Web3NetAPI>;
+			privateKey?: HexString | Uint8Array;
+			fillGasPrice?: boolean;
+			fillGasLimit?: boolean;
+		}): Promise<ReturnType> {
+			const tx = { ...options.transaction };
+			return tx as unknown as ReturnType;
+		}
+
+		context.transactionBuilder = transactionBuilder;
+
+		const ethereumjsTx = await prepareTransactionForSigning(
+			{
+				chainId: 1458,
+				nonce: 1,
+				gasPrice: BigInt(20000000000),
+				gas: BigInt(21000),
+				to: '0xF0109fC8DF283027b6285cc889F5aA624EaC1F55',
+				from: '0x2c7536E3605D9C16a7a3D7b1898e529396a65c23',
+				value: '1000000000',
+				input: '',
+				networkId: 999,
+			},
+			context,
+		);
+		expect(ethereumjsTx.common.hardfork()).toBe(Hardfork.Istanbul);
+		expect(ethereumjsTx.common.networkId().toString()).toBe('999');
+	});
+
+	it('should be able to read Hardfork from context.config.defaultHardfork and context.defaultCommon.hardfork', async () => {
+		const context = new Web3Context<EthExecutionAPI>({
+			provider: new HttpProvider('http://127.0.0.1'),
+			config: { defaultNetworkId: '0x9' },
+		});
+		context.defaultChain = 'mainnet';
+
+		// if the value here is different from the one in context.defaultCommon.hardfork
+		// Then an error will be thrown:
+		// "ConfigHardforkMismatchError: Web3Config hardfork doesnt match in defaultHardfork london and common.hardfork istanbul"
+		context.config.defaultHardfork = Hardfork.Istanbul;
+		context.defaultCommon = {
+			customChain: {
+				name: 'test',
+				networkId: 111,
+				chainId: 1458,
+			},
+			hardfork: Hardfork.Istanbul,
+			baseChain: 'mainnet',
+		} as any;
+
+		async function transactionBuilder<ReturnType = TransactionType>(options: {
+			transaction: TransactionType;
+			web3Context: Web3Context<EthExecutionAPI & Web3NetAPI>;
+			privateKey?: HexString | Uint8Array;
+			fillGasPrice?: boolean;
+			fillGasLimit?: boolean;
+		}): Promise<ReturnType> {
+			const tx = { ...options.transaction };
+			return tx as unknown as ReturnType;
+		}
+
+		context.transactionBuilder = transactionBuilder;
+
+		const ethereumjsTx = await prepareTransactionForSigning(
+			{
+				chainId: 1458,
+				nonce: 1,
+				gasPrice: BigInt(20000000000),
+				gas: BigInt(21000),
+				to: '0xF0109fC8DF283027b6285cc889F5aA624EaC1F55',
+				from: '0x2c7536E3605D9C16a7a3D7b1898e529396a65c23',
+				value: '1000000000',
+				input: '',
+			},
+			context,
+		);
+		expect(ethereumjsTx.common.hardfork()).toBe(Hardfork.Istanbul);
+		expect(ethereumjsTx.common.networkId().toString()).toBe('111');
+	});
+
+	it('should give priorities to tx.hardfork and tx.networkId over values from context', async () => {
+		const context = new Web3Context<EthExecutionAPI>({
+			provider: new HttpProvider('http://127.0.0.1'),
+			config: { defaultNetworkId: '0x9' },
+		});
+		context.defaultChain = 'mainnet';
+
+		// if the value here is different from the one in context.defaultCommon.hardfork
+		// Then an error will be thrown:
+		// "ConfigHardforkMismatchError: Web3Config hardfork doesnt match in defaultHardfork london and common.hardfork istanbul"
+		context.config.defaultHardfork = Hardfork.Istanbul;
+		context.defaultCommon = {
+			customChain: {
+				name: 'test',
+				networkId: 111,
+				chainId: 1458,
+			},
+			hardfork: Hardfork.Istanbul,
+			baseChain: 'mainnet',
+		} as any;
+
+		async function transactionBuilder<ReturnType = TransactionType>(options: {
+			transaction: TransactionType;
+			web3Context: Web3Context<EthExecutionAPI & Web3NetAPI>;
+			privateKey?: HexString | Uint8Array;
+			fillGasPrice?: boolean;
+			fillGasLimit?: boolean;
+		}): Promise<ReturnType> {
+			const tx = { ...options.transaction };
+			return tx as unknown as ReturnType;
+		}
+
+		context.transactionBuilder = transactionBuilder;
+
+		// context.transactionBuilder = defaultTransactionBuilder;
+
+		const ethereumjsTx = await prepareTransactionForSigning(
+			{
+				chainId: 1458,
+				nonce: 1,
+				gasPrice: BigInt(20000000000),
+				gas: BigInt(21000),
+				to: '0xF0109fC8DF283027b6285cc889F5aA624EaC1F55',
+				from: '0x2c7536E3605D9C16a7a3D7b1898e529396a65c23',
+				value: '1000000000',
+				input: '',
+				networkId: 999,
+				hardfork: Hardfork.Chainstart,
+				chain: 'mainnet',
+			},
+			context,
+		);
+		expect(ethereumjsTx.common.hardfork()).toBe(Hardfork.Chainstart);
+		expect(ethereumjsTx.common.networkId().toString()).toBe('999');
+	});
+
 	describe('should return an web3-utils/tx instance with expected properties', () => {
 		it.each(validTransactions)(
 			'mockBlock: %s\nexpectedTransaction: %s\nexpectedPrivateKey: %s\nexpectedAddress: %s\nexpectedRlpEncodedTransaction: %s\nexpectedTransactionHash: %s\nexpectedMessageToSign: %s\nexpectedV: %s\nexpectedR: %s\nexpectedS: %s',