+++ packages/web3-eth-accounts/CHANGELOG.md
@@ -150,3 +150,8 @@ Documentation:
 -   Fixed `recover` function, `v` will be normalized to value 0,1 (#6344) 
 
 ## [Unreleased]
+
+### Fixed
+
+-   Send Transaction config used to be ignored if the passed `common` did not have a `copy()` and the `chainId` was not provided (#6663)
+-   Fixed an issue with detecting Uint8Array (#6486)